fix(rope): bf16 cos/sin precision + partial_rotary guard; bump polygram 0.10
compute_rope_cache honored dtype=q.dtype, so on M4 bf16/MPS it built
inv_freq, the position grid, and cos/sin entirely in bf16. At Gemma-2's
head_dim=256, arange(seq_len, dtype=bf16) aliases integer positions above
256 (bf16's 7-bit mantissa), and the smallest inv_freq (~1e-4) loses ~half
its precision. Measured cos drift up to 2.0 (the full range, saturated) at
seq_len >= 512 — i.e. randomized attention pattern, not small numerical
error. Pre-softmax attention-logit relative drift was 4.5% at seq_len=64,
41% at 512, 91% at 8192.

Match HF convention: build inv_freq, t, cos/sin in fp32 unconditionally,
cast to caller dtype at the end. End-to-end forge-vs-HF Llama on identity
basis now matches host to float-noise at fp32 (~1e-7) and to bf16
activation-noise at bf16 (~5e-3, constant in seq_len — the smoking gun
that the fix removed the seq-len-dependent drift).

partial_rotary_factor was read into NativeModelConfig but never used in
LlamaSelfAttention.forward; Gemma-2 / Llama / Qwen2-3 all use 1.0 so this
was silently fine for them, but Gemma-3 / GPT-J / NeoX-style hosts with
factor under 1.0 would have rotated all head_dim instead of the partial
slice. Now raises NotImplementedError at first forward, parallel to the
existing rope_scaling.type guard.

Also bumps the polygram pin to >=0.10.0 — promotes EpochReport.redundancy_ratio
and n_features_input to first-class fields (schema v2 with v1 forward-compat
shim) and emits a UserWarning on BlockFormation degenerate partitions.

Tests: 627 passed, 10 skipped. New regression coverage:
- test_compute_rope_cache_bf16_matches_fp32_cast: pins bit-equality against
  the fp32-then-cast path on Gemma-2 head_dim=256, seq_len=512.
- test_partial_rotary_factor_non_unity_raises_at_forward: pins that a forged
  module with factor=0.5 raises at first forward.

+++ saeforge/_positional/rope.py
@@ -56,15 +56,26 @@ def compute_rope_cache(
     torch = require_extra("torch", "torch")
     if dtype is None:
         dtype = torch.float32
+    # HF convention: build inv_freq, the position grid, and cos/sin in
+    # fp32 unconditionally, then cast the cache to `dtype` at the end.
+    # Honoring `dtype=q.dtype` (the previous behavior) breaks badly at
+    # bf16: `torch.arange(seq_len, dtype=bf16)` aliases integer
+    # positions above 256 (bf16's 7-bit mantissa), and at Gemma-2's
+    # head_dim=256 the smallest inv_freq (~1e-4) loses ~half its
+    # precision. Measured cos drift up to 2.0 (full range) at
+    # seq_len>=512 with q.dtype=bf16 — i.e. attention pattern noise,
+    # not a small numerical error. fp32 here costs ~head_dim*seq_len*4
+    # bytes per forward (a few MB at production scale).
     inv_freq = 1.0 / (
         theta ** (
-            torch.arange(0, head_dim, 2, device=device, dtype=dtype) / head_dim
+            torch.arange(0, head_dim, 2, device=device, dtype=torch.float32)
+            / head_dim
         )
     )
-    t = torch.arange(seq_len, device=device, dtype=dtype)
+    t = torch.arange(seq_len, device=device, dtype=torch.float32)
     freqs = torch.outer(t, inv_freq)  # (seq_len, head_dim/2)
     emb = torch.cat((freqs, freqs), dim=-1)  # (seq_len, head_dim)
-    return emb.cos(), emb.sin()
+    return emb.cos().to(dtype), emb.sin().to(dtype)
 
 
 def _rotate_half(x):
